vehicle-maintenance-agent/backend/requirements.txt
#utils
requests
python-dateutil
python-dotenv
uvicorn
fastapi
ibm-watsonx-orchestrate

# Optional visualization
# plotly

# pulp

